--- src/components/Tab.module.scss
@@ -1,13 +1,0 @@
-:local {
-  .tab {
-    cursor: pointer;
-    background-color: #f3f3f3;
-    border: 1px solid #bbb;
-    border-radius: 3px 3px 0 0;
-    display: inline-block;
-    margin-bottom: -1px;
-    margin-left: 1px;
-    padding: 3px 5px;
-    pointer-events: all;
-  }
-}
