--- src/components/Tabs.js
@@ -1,13 +1,0 @@
-import jsx from '../utilities/jsx';
-import Tab from './Tab';
-
-/**
- * Tab bar with content
- */
-export default ({ console, setContent }) => {
-  return (
-    <>
-      <Tab label='Console' onclick={() => setContent(console)}/>
-    </>
-  );
-};
