+++ src/components/tabs/console/Output.module.scss
