+++ src/components/tabs/console/Prompt.module.scss
