+++ src/components/tabs/console/Prompt.test.js
